+++ src/EntityGraphQL/Directives/DirectiveProcessor.cs
@@ -1,7 +1,7 @@
 using System;
 using System.Collections.Generic;
 using System.Linq;
-using EntityGraphQL.Compiler;
+using System.Linq.Expressions;
 using EntityGraphQL.Schema;
 
 namespace EntityGraphQL.Directives
@@ -15,44 +15,19 @@ namespace EntityGraphQL.Directives
     public abstract class DirectiveProcessor<TArguments> : IDirectiveProcessor
     {
         public abstract Type GetArgumentsType();
-        public abstract bool ProcessesResult { get; }
         public abstract string Name { get; }
         public abstract string Description { get; }
 
         /// <summary>
-        /// Implement this to make changes to the item marked with the directive.
-        /// GraphQLQueryNode details the expressions for the field query
-        /// </summary>
-        /// <param name="field"></param>
-        /// <param name="arguments">Any arguments passed to the directive</param>
-        /// <returns>Return a modified (from field) or new IGraphQLBaseNode. Returning null will remove the item from the resulting query</returns>
-        public virtual BaseGraphQLField? ProcessQuery(BaseGraphQLField field, TArguments arguments)
-        {
-            // by default we do nothing
-            return field;
-        }
-
-        /// <summary>
-        /// Implement this to make changes to the result
+        /// Implement this to make changes to the expression that will execute
         /// </summary>
         /// <param name="value"></param>
         /// <param name="arguments"></param>
         /// <returns></returns>
-        public virtual object ProcessResult(object value, TArguments arguments)
+        public virtual Expression? ProcessExpression(Expression expression, object arguments)
         {
             // default return the value
-            return value;
-        }
-
-        public BaseGraphQLField? ProcessField(BaseGraphQLField field, object arguments)
-        {
-            var result = ProcessQuery(field, (TArguments)arguments);
-            if (ProcessesResult)
-            {
-                // wrap expression in a call to process the result
-                // field.Wrap(exp => Expression.Call(Expression.Constant(this), "ProcessResult", null, exp, arguments));
-            }
-            return result;
+            return expression;
         }
 
         public IEnumerable<ArgType> GetArguments(ISchemaProvider schema, Func<string, string> fieldNamer)