+++ src/EntityGraphQL/Directives/IDirectiveProcessor.cs
@@ -1,5 +1,6 @@
 using System;
 using System.Collections.Generic;
+using System.Linq.Expressions;
 using EntityGraphQL.Compiler;
 using EntityGraphQL.Schema;
 
@@ -14,12 +15,7 @@ namespace EntityGraphQL.Directives
         /// </summary>
         /// <returns></returns>
         Type GetArgumentsType();
-        /// <summary>
-        /// Return true if the directive requires to make changes to the result
-        /// </summary>
-        /// <value></value>
-        bool ProcessesResult { get; }
-        BaseGraphQLField? ProcessField(BaseGraphQLField fieldResult, object arguments);
+        Expression? ProcessExpression(Expression expression, object arguments);
         IEnumerable<ArgType> GetArguments(ISchemaProvider schema, Func<string, string> fieldNamer);
     }
 }