+++ src/EntityGraphQL/Directives/SkipDirectiveProcessor.cs
@@ -1,6 +1,6 @@
 using System;
 using System.ComponentModel;
-using EntityGraphQL.Compiler;
+using System.Linq.Expressions;
 
 namespace EntityGraphQL.Directives
 {
@@ -8,17 +8,16 @@ namespace EntityGraphQL.Directives
     {
         public override string Name { get => "skip"; }
         public override string Description { get => "Directs the executor to skip this field or fragment when the `if` argument is true."; }
-        public override bool ProcessesResult { get => false; }
         public override Type GetArgumentsType()
         {
             return typeof(SkipArguments);
         }
 
-        public override BaseGraphQLField? ProcessQuery(BaseGraphQLField field, SkipArguments arguments)
+        public override Expression? ProcessExpression(Expression expression, object arguments)
         {
-            if (arguments.@if)
+            if (((SkipArguments)arguments).@if)
                 return null;
-            return field;
+            return expression;
         }
     }